fix: set enableReusePort on socket so bind() doesn't fail
There seems to be a race between close() and subsequent bind()'s.
This can be seen when a deploy redirects or when calling start() back to
back (any time there is an active socket and we close it immediately
before creating a new one).

We could also possibly not create a new socket every time on start().